Move a suite runner into its own package so it can be renamed
This may be not entirely desirable, but it moves the last *Test class
that purely aggregates test classes into a suite to a package where it
can be named PackageTest.
